テキスト抽出系を再配置。単純処理だけなので同じファイルで大丈夫でしょう
