+++ src/rmmz/plugin/core/params/index.ts
@@ -1,5 +1,4 @@
 export * from "./arraySchemaFilterByParam";
-export * from "./arraySchemaFilterByParamEx";
 export * from "./classify";
 export * from "./convert";
 export * from "./lookupRpgDataKind";
